services/api-service/requirements.txt
fastapi
uvicorn[standard]
redis
grpcio
grpcio-tools
protobuf
kafka-python-ng